fix: tiny change

+++ src/code/math/euler_phi_u64.hpp
@@ -7,7 +7,7 @@
 namespace tifa_libs::math {
 
 inline u64 euler_phi_u64(u64 n) {
-  if (n <= (u32)(-1)) return euler_phi_u32((u32)n);
+  if (n <= (u64)-1_u32) return euler_phi_u32((u32)n);
   auto pf = pfactors(n);
   for (auto [p, k] : pf) n = n / p * (p - 1);
   return n;
